requirements.txt
simplekml
numpy
Pillow
opencv-python
exifread
flask
joblib